+++ examples/MinimalInterrupt/MinimalInterrupt.ino
@@ -94,7 +94,7 @@ void setup(){
 
     rfm.setBufferSize(2);   // set the internal buffer size.
     rfm.setPacketLength(4); // set the packet length.
-    rfm.setFrequency(434*1000*1000); // set the frequency.
+    rfm.setFrequency((uint32_t) 434*1000*1000); // set the frequency.
 
     // baudrate is default, 4800 bps now.
